backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.2
sqlalchemy==2.0.48
flask-sqlalchemy==3.1.1
passlib==1.7.4
PyJWT==2.12.1
